slotTool 1.4
This slotTool version doesn't need special otherapps with cfg:s added. It uses a process takeover .xml to get mset's cfg:i in order to install usm. Note the change that you need to run this in a folder with the provided .xml file (at least in a userland context).

This version is also entrypoint agnostic, it will even work
with exploits that don't use otherapp (ex. ninjhax).